+++ src/components/SectionTitle.vue
@@ -1,11 +1,20 @@
 <script setup lang="ts">
-import { onMounted } from 'vue'
+import { computed, type ComputedRef, type PropType, onMounted } from 'vue'
 
+import { Stability } from '@/types/enum'
 import type { TocItem } from '@/types/app'
 
 import Link from '@/components/Link.vue'
 
 const props = defineProps({
+  version: {
+    type: String,
+    required: true
+  },
+  stability: {
+    type: String as PropType<Stability>,
+    default: Stability.Stable
+  },
   anchor: {
     type: String,
     required: true
@@ -31,6 +40,18 @@ const tocItem: TocItem = {
 
 const emit = defineEmits(['update:tocItems'])
 
+let stabilityClasses: ComputedRef<string[]> = computed(() => {
+  const classes = []
+
+  if (props.stability === Stability.Experimental) {
+    classes.push('text-purple-500')
+  } else if (props.stability === Stability.Stable) {
+    classes.push('text-green-500')
+  }
+
+  return classes
+})
+
 onMounted(() => {
   emit('update:tocItems', tocItem)
 })
@@ -39,9 +60,13 @@ onMounted(() => {
 <template>
   <header class="mb-5 space-y-2">
     <div class="flex justify-between items-center">
-      <h2 :id="anchor" class="text-2xl font-semibold flex-grow">
-        {{ title }}
-      </h2>
+      <div class="flex space-x-2">
+        <h2 :id="anchor" class="text-2xl font-semibold flex-grow">{{ title }}</h2>
+        <div class="flex flex-col justify-center">
+          <div class="section-version text-[8px] text-neutral-500">{{ version }}</div>
+          <div class="section-stability text-[8px]" :class="stabilityClasses">{{ stability }}</div>
+        </div>
+      </div>
       <div class="flex gap-4">
         <Link class="section-guidance" v-if="guidanceHref" :href="guidanceHref">View Guidance</Link>
         <Link